+++ CHANGELOG.md
@@ -12,3 +12,22 @@ This is our initial release of the plugin
 
 ### Fixes
 - Fixed Looping not working of FireFox
+
+## [0.0.3] Initial Fixes - 2022-07-03
+
+### Added
+- Touch events 
+- Event on frame change "wpe/ScrollSequence/frame"
+
+#### "wpe/ScrollSequence/frame"
+Every time the frame changes this event is dispatched on window.
+The Container ID and teh current frame are in ethe event detail
+
+eg.
+<pre>
+window.addEventListener('wpe/ScrollSequence/frame', e => { 
+    const {id, frame} = e.detail;
+    console.log('id': id);
+    console.log('frame': frame);
+})
+</pre>